filter fixed

+++ app/src/main/java/ch/zhaw/bait17/audio_signal_processing_toolbox/dsp/filter/Filter.java
@@ -19,7 +19,7 @@ public interface Filter extends Parcelable {
      * @param input
      * @return
      */
-    short[] apply(@NonNull short[] input);
+    float[] apply(@NonNull float[] input);
 
     FilterSpec getFilterSpec();
